formatting

+++ contracts/MoatPayment.sol
@@ -7,47 +7,69 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 //import "https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/token/ERC20/IERC20.sol";
 
 contract MoatPayment {
-
-    event MoatCreated(string indexed moatName, address indexed validator, address indexed creator);
-    event MoatFunded(string indexed moatName, address indexed validator, address indexed creator, uint fundsAdded, uint totalFunds);
-    event ValidatorTransferred(string indexed moatName, address indexed newValidator, address indexed oldValidator);
+    event MoatCreated(
+        string indexed moatName,
+        address indexed validator,
+        address indexed creator
+    );
+    event MoatFunded(
+        string indexed moatName,
+        address indexed validator,
+        address indexed creator,
+        uint256 fundsAdded,
+        uint256 totalFunds
+    );
+    event ValidatorTransferred(
+        string indexed moatName,
+        address indexed newValidator,
+        address indexed oldValidator
+    );
 
     IERC20 public escrowToken;
 
     struct Moat {
-        uint pool;
+        uint256 pool;
         address validator;
         address creator;
     }
 
-    mapping (string => Moat) public moats;
+    mapping(string => Moat) public moats;
 
     constructor(address _escrowToken) {
         escrowToken = IERC20(_escrowToken);
     }
 
     modifier isValidator(string calldata moatName) {
-        require(moats[moatName].validator == msg.sender, "MoatPayment: you are not the validator for this moat.");
+        require(
+            moats[moatName].validator == msg.sender,
+            "MoatPayment: you are not the validator for this moat."
+        );
         _;
     }
 
     /**
      *  Creates a new moat (pool). Will fail if another moat of the same name has been registered in the past.
      */
-    function createMoat (string calldata moatName, address validator) public {
-		require(moats[moatName].creator == address(0x0), "MoatPayment: moat already exists.");
-		moats[moatName] = Moat(0, validator, msg.sender);
+    function createMoat(string calldata moatName, address validator) public {
+        require(
+            moats[moatName].creator == address(0x0),
+            "MoatPayment: moat already exists."
+        );
+        moats[moatName] = Moat(0, validator, msg.sender);
 
         emit MoatCreated(moatName, validator, msg.sender);
-	}
+    }
 
     /**
      *  Adds fund to a moat.
      */
-    function fundMoat (string calldata moatName, uint amt) public payable {
+    function fundMoat(string calldata moatName, uint256 amt) public payable {
         Moat storage moat = moats[moatName];
-        require(moat.validator != address(0x0), "MoatPayment: moat does not exist.");
-        
+        require(
+            moat.validator != address(0x0),
+            "MoatPayment: moat does not exist."
+        );
+
         /** NOTE:
          *  A smart contract cannot approve tokens for someone by design (security).
          *  This is what has to happen in your frontend:
@@ -65,9 +87,15 @@ contract MoatPayment {
     /**
      *  Allows the validator to withdraw a token from a moat (pool).
      */
-    function withdrawFromPool (string calldata moatName, uint amt) public isValidator(moatName) {
+    function withdrawFromPool(string calldata moatName, uint256 amt)
+        public
+        isValidator(moatName)
+    {
         Moat storage moat = moats[moatName];
-        require(moat.pool >= amt, "MoatPayment: not enough value in moat's pool.");
+        require(
+            moat.pool >= amt,
+            "MoatPayment: not enough value in moat's pool."
+        );
 
         moats[moatName].pool -= amt;
         bool success = escrowToken.transfer(moat.validator, amt);
@@ -77,11 +105,16 @@ contract MoatPayment {
     /**
      * Transfers the validator role of a moat to another address.
      */
-    function transferValidator (string calldata moatName, address newValidator) public isValidator(moatName) {
-		require(newValidator != address(0x0), "MoatPayment: do not transfer validator to the zero address.");
+    function transferValidator(string calldata moatName, address newValidator)
+        public
+        isValidator(moatName)
+    {
+        require(
+            newValidator != address(0x0),
+            "MoatPayment: do not transfer validator to the zero address."
+        );
         address oldValidator = moats[moatName].validator;
         moats[moatName].validator = newValidator;
         emit ValidatorTransferred(moatName, newValidator, oldValidator);
     }
-
 }
